inference/pyproject.toml
[project]
name = "support-copilot-inference"
version = "0.1.0"
description = "Inference pipeline: hybrid retrieval, reranking, generation, grading"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform>=1.74.0",
    "google-cloud-bigquery>=3.27.0",
    "google-genai>=1.47.0",
    "jinja2>=3.1.0",
    "numpy>=1.26.0",
    "openai>=1.58.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.0",
    "sentence-transformers>=3.3.0",
    "wandb>=0.19.0",
    "pydantic-settings>=2.7.0",
]
